+++ Tests/LinuxMain.swift
@@ -113,6 +113,16 @@ extension ConditionalReturnsOnNewlineRuleTests {
     ]
 }
 
+extension ConfigurationAliasesTests {
+    static var allTests: [(String, (ConfigurationAliasesTests) -> () throws -> Void)] = [
+        ("testConfiguresCorrectlyFromDeprecatedAlias", testConfiguresCorrectlyFromDeprecatedAlias),
+        ("testReturnsNilWithDuplicatedConfiguration", testReturnsNilWithDuplicatedConfiguration),
+        ("testInitsFromDeprecatedAlias", testInitsFromDeprecatedAlias),
+        ("testWhitelistRulesFromDeprecatedAlias", testWhitelistRulesFromDeprecatedAlias),
+        ("testDisabledRulesFromDeprecatedAlias", testDisabledRulesFromDeprecatedAlias)
+    ]
+}
+
 extension ConfigurationTests {
     static var allTests: [(String, (ConfigurationTests) -> () throws -> Void)] = [
         ("testInit", testInit),
@@ -132,6 +142,7 @@ extension ConfigurationTests {
         ("testForceExcludesDirectory", testForceExcludesDirectory),
         ("testForceExcludesDirectoryThatIsNotInExcludedButHasChildrenThatAre", testForceExcludesDirectoryThatIsNotInExcludedButHasChildrenThatAre),
         ("testLintablePaths", testLintablePaths),
+        ("testGlobExcludePaths", testGlobExcludePaths),
         ("testIsEqualTo", testIsEqualTo),
         ("testIsNotEqualTo", testIsNotEqualTo),
         ("testCustomConfiguration", testCustomConfiguration),
@@ -142,11 +153,6 @@ extension ConfigurationTests {
         ("testIndentationFallback", testIndentationFallback),
         ("testConfiguresCorrectlyFromDict", testConfiguresCorrectlyFromDict),
         ("testConfigureFallsBackCorrectly", testConfigureFallsBackCorrectly),
-        ("testConfiguresCorrectlyFromDeprecatedAlias", testConfiguresCorrectlyFromDeprecatedAlias),
-        ("testReturnsNilWithDuplicatedConfiguration", testReturnsNilWithDuplicatedConfiguration),
-        ("testInitsFromDeprecatedAlias", testInitsFromDeprecatedAlias),
-        ("testWhitelistRulesFromDeprecatedAlias", testWhitelistRulesFromDeprecatedAlias),
-        ("testDisabledRulesFromDeprecatedAlias", testDisabledRulesFromDeprecatedAlias),
         ("testMerge", testMerge),
         ("testLevel0", testLevel0),
         ("testLevel1", testLevel1),
@@ -465,6 +471,16 @@ extension GenericTypeNameRuleTests {
         ("testGenericTypeNameWithAllowedSymbols", testGenericTypeNameWithAllowedSymbols),
         ("testGenericTypeNameWithAllowedSymbolsAndViolation", testGenericTypeNameWithAllowedSymbolsAndViolation),
         ("testGenericTypeNameWithIgnoreStartWithLowercase", testGenericTypeNameWithIgnoreStartWithLowercase)
+    ]
+}
+
+extension GlobTests {
+    static var allTests: [(String, (GlobTests) -> () throws -> Void)] = [
+        ("testOnlyGlobForWildcard", testOnlyGlobForWildcard),
+        ("testNoMatchReturnsEmpty", testNoMatchReturnsEmpty),
+        ("testMatchesFiles", testMatchesFiles),
+        ("testMatchesMultipleFiles", testMatchesMultipleFiles),
+        ("testMatchesNestedDirectory", testMatchesNestedDirectory)
     ]
 }
 
@@ -1201,6 +1217,7 @@ XCTMain([
     testCase(CommandTests.allTests),
     testCase(CompilerProtocolInitRuleTests.allTests),
     testCase(ConditionalReturnsOnNewlineRuleTests.allTests),
+    testCase(ConfigurationAliasesTests.allTests),
     testCase(ConfigurationTests.allTests),
     testCase(ContainsOverFirstNotNilRuleTests.allTests),
     testCase(ControlStatementRuleTests.allTests),
@@ -1244,6 +1261,7 @@ XCTMain([
     testCase(FunctionDefaultParameterAtEndRuleTests.allTests),
     testCase(FunctionParameterCountRuleTests.allTests),
     testCase(GenericTypeNameRuleTests.allTests),
+    testCase(GlobTests.allTests),
     testCase(IdentifierNameRuleTests.allTests),
     testCase(ImplicitGetterRuleTests.allTests),
     testCase(ImplicitReturnRuleTests.allTests),