+++ CHANGELOG.md
@@ -37,6 +37,10 @@
   [Cihat Gündüz](https://github.com/Dschee)
   [#2307](https://github.com/realm/SwiftLint/issues/2307)
 
+* Add support for globs in `excluded` file paths  
+  [Keith Smiley](https://github.com/keith)
+  [#2316](https://github.com/realm/SwiftLint/pull/2316)
+
 * Add `private_only` configuration to `prefixed_toplevel_constant` rule  
   [Keith Smiley](https://github.com/keith)
   [#2315](https://github.com/realm/SwiftLint/pull/2315)